+++ flexmeasures/api/v3_0/tests/conftest.py
@@ -2,12 +2,14 @@ from __future__ import annotations
 
 from datetime import timedelta
 
+import numpy as np
 import pandas as pd
 import pytest
 from flask_security import SQLAlchemySessionUserDatastore, hash_password
 from sqlalchemy import select, delete
 
 from flexmeasures import Sensor, Source, User, UserRole
+from flexmeasures.data.models.data_sources import DataSource
 from flexmeasures.data.models.generic_assets import GenericAssetType, GenericAsset
 from flexmeasures.data.models.time_series import TimedBelief
 
@@ -150,6 +152,11 @@ def add_incineration_line(db, test_supplier_user) -> dict[str, Sensor]:
     )
     db.session.add(gas_sensor)
     add_gas_measurements(db, test_supplier_user.data_source[0], gas_sensor)
+    other_source = DataSource(name="Other source", type="demo script")
+    db.session.add(other_source)
+    db.session.flush()
+    add_gas_measurements(db, other_source, gas_sensor, values=[91.3, np.nan, 92.1])
+
     temperature_sensor = Sensor(
         name="some temperature sensor",
         unit="°C",
@@ -177,12 +184,12 @@ def add_incineration_line(db, test_supplier_user) -> dict[str, Sensor]:
     }
 
 
-def add_gas_measurements(db, source: Source, sensor: Sensor):
+def add_gas_measurements(db, source: Source, sensor: Sensor, values=None):
     event_starts = [
         pd.Timestamp("2021-05-02T00:00:00+02:00") + timedelta(minutes=minutes)
         for minutes in range(0, 30, 10)
     ]
-    event_values = [91.3, 91.7, 92.1]
+    event_values = list(values) if values else [91.3, 91.7, 92.1]
     beliefs = [
         TimedBelief(
             sensor=sensor,