+++ flexmeasures/ui/utils/view_utils.py
@@ -1,4 +1,5 @@
 """Utilities for views"""
+
 from __future__ import annotations
 
 import json
@@ -86,14 +87,24 @@ def render_flexmeasures_template(html_filename: str, **variables):
         options["downloadFileName"] = f"asset-{asset.id}-{asset.name}"
     variables["chart_options"] = json.dumps(options)
 
-    variables["menu_logo"] = current_app.config.get("FLEXMEASURES_MENU_LOGO_PATH")
+    account: Account | None = (
+        current_user.account if current_user.is_authenticated else None
+    )
+
+    # check if user/consultant has logo_url set
+    if account:
+        variables["menu_logo"] = (
+            account.logo_url
+            or (account.consultancy_account and account.consultancy_account.logo_url)
+            or current_app.config.get("FLEXMEASURES_MENU_LOGO_PATH")
+        )
+    else:
+        variables["menu_logo"] = current_app.config.get("FLEXMEASURES_MENU_LOGO_PATH")
+
     variables["extra_css"] = current_app.config.get("FLEXMEASURES_EXTRA_CSS_PATH")
 
     if "asset" in variables:
         variables["breadcrumb_info"] = get_breadcrumb_info(asset)
-    account: Account | None = (
-        current_user.account if current_user.is_authenticated else None
-    )
     variables.update(get_color_settings(account))  # add color settings to variables
 
     return render_template(html_filename, **variables)
@@ -166,6 +177,23 @@ def get_git_description() -> tuple[str, int, str]:
     return version, commits_since, sha
 
 
+ICON_MAPPING = {
+    # site structure
+    "evse": "icon-charging_station",
+    "charge point": "icon-charging_station",
+    "project": "icon-calculator",
+    "tariff": "icon-time",
+    "renewables": "icon-wind",
+    "site": "icon-empty-marker",
+    "scenario": "icon-binoculars",
+    # weather
+    "irradiance": "wi wi-horizon-alt",
+    "temperature": "wi wi-thermometer",
+    "wind direction": "wi wi-wind-direction",
+    "wind speed": "wi wi-strong-wind",
+}
+
+
 def asset_icon_name(asset_type_name: str) -> str:
     """Icon name for this asset type.
 
@@ -177,27 +205,9 @@ def asset_icon_name(asset_type_name: str) -> str:
     becomes (for a battery):
         <i class="icon-battery"></i>
     """
-    icon_mapping = {
-        # site structure
-        "evse": "icon-charging_station",
-        "charge point": "icon-charging_station",
-        "project": "icon-calculator",
-        "tariff": "icon-time",
-        "renewables": "icon-wind",
-        "site": "icon-empty-marker",
-        "scenario": "icon-binoculars",
-        # weather
-        "irradiance": "wi wi-horizon-alt",
-        "temperature": "wi wi-thermometer",
-        "wind direction": "wi wi-wind-direction",
-        "wind speed": "wi wi-strong-wind",
-    }
-
-    for asset_group_name, icon_name in icon_mapping.items():
-        if asset_group_name in asset_type_name.lower():
-            return icon_name
-
-    return f"icon-{asset_type_name}"
+    if asset_type_name:
+        asset_type_name = asset_type_name.lower()
+    return ICON_MAPPING.get(asset_type_name, f"icon-{asset_type_name}")
 
 
 def username(user_id) -> str: